# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ProjetoZup API",
    "version": "1.0.0",
    "description": "API REST para gestão de infraestrutura urbana (categorias, subcategorias, bairros e ocorrências com geolocalização).",
    "license": { "name": "GPL-3.0" }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
...
  "paths": {
...
    "/api/categories": {
      "get": {
        "tags": ["Categories"],
        "summary": "Listar categorias",
        "operationId": "listCategories",
        "responses": {
          "200": {
            "description": "Lista de categorias",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Category" } }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Categories"],
        "summary": "Criar categoria (admin)",
        "description": "Requer autenticação e papel admin.",
        "operationId": "createCategory",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CategoryCreateInput" },
              "example": {
                "name": "Iluminação pública",
                "slug": "iluminacao-publica",
                "description": "Problemas relacionados à iluminação de vias e espaços públicos",
                "icon": "lightbulb",
                "color": "#FFC107",
                "is_active": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Categoria criada",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Category" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/categories/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
      "get": {
        "tags": ["Categories"],
        "summary": "Obter categoria por id",
        "operationId": "getCategoryById",
        "responses": {
          "200": {
            "description": "Categoria",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Category" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["Categories"],
        "summary": "Atualizar categoria (admin)",
        "description": "Atualização parcial. Requer autenticação e papel admin.",
        "operationId": "updateCategory",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CategoryUpdateInput" },
              "example": { "icon": "wb_incandescent", "color": "#FFB300" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Categoria atualizada",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Category" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      },
      "delete": {
        "tags": ["Categories"],
        "summary": "Remover categoria (admin)",
        "description": "Requer autenticação e papel admin.",
        "operationId": "deleteCategory",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Removida" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/subcategories": {
      "get": {
        "tags": ["Subcategories"],
        "summary": "Listar subcategorias",
        "operationId": "listSubcategories",
        "responses": {
          "200": {
            "description": "Lista de subcategorias",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Subcategory" } }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Subcategories"],
        "summary": "Criar subcategoria (admin)",
        "description": "Requer autenticação e papel admin.",
        "operationId": "createSubcategory",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SubcategoryCreateInput" },
              "example": {
                "category_id": 3,
                "name": "Poste queimado",
                "slug": "poste-queimado",
                "description": "Poste de iluminação com lâmpada queimada",
                "icon": "lightbulb",
                "is_active": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Subcategoria criada",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Subcategory" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/subcategories/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
      "get": {
        "tags": ["Subcategories"],
        "summary": "Obter subcategoria por id",
        "operationId": "getSubcategoryById",
        "responses": {
          "200": {
            "description": "Subcategoria",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Subcategory" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["Subcategories"],
        "summary": "Atualizar subcategoria (admin)",
        "description": "Atualização parcial. Requer autenticação e papel admin. Trocar category_id/name/slug revalida a consistência (categoria existente; nome/slug únicos na categoria-alvo).",
        "operationId": "updateSubcategory",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SubcategoryUpdateInput" },
              "example": { "icon": "wb_incandescent", "is_active": false }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subcategoria atualizada",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Subcategory" } }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      },
      "delete": {
        "tags": ["Subcategories"],
        "summary": "Remover subcategoria (admin)",
        "description": "Requer autenticação e papel admin.",
        "operationId": "deleteSubcategory",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Removida" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT emitido por /api/auth/login ou /api/auth/refresh. Use o header `Authorization: Bearer <access_token>`."
      }
    },
    "parameters": {
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "minimum": 1 },
        "example": 1
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Requisição inválida",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Unauthorized": {
        "description": "Não autenticado ou token inválido/expirado",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Forbidden": {
        "description": "Usuário autenticado sem permissão para o recurso",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NotFound": {
        "description": "Recurso não encontrado",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Conflict": {
        "description": "Conflito de estado",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" },
          "details": { "type": "object", "additionalProperties": true, "nullable": true }
        }
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string", "maxLength": 100 },
          "slug": { "type": "string", "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$", "maxLength": 120, "description": "Chave textual estável para uso no front (única globalmente)." },
          "description": { "type": "string", "nullable": true },
          "icon": { "type": "string", "maxLength": 50, "nullable": true, "description": "Identificador nominal do ícone (ex.: nome em uma biblioteca de ícones do front). Não é um arquivo." },
          "color": { "type": "string", "pattern": "^#[0-9a-fA-F]{6}$", "nullable": true },
          "is_active": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "CategoryCreateInput": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "maxLength": 100 },
          "slug": { "type": "string", "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$", "maxLength": 120, "description": "Opcional. Se omitido, é derivado do nome (sem acentos)." },
          "description": { "type": "string" },
          "icon": { "type": "string", "maxLength": 50 },
          "color": { "type": "string", "pattern": "^#[0-9a-fA-F]{6}$" },
          "is_active": { "type": "boolean", "default": true }
        }
      },
      "CategoryUpdateInput": {
        "type": "object",
        "description": "Atualização parcial: envie apenas os campos a alterar. description/icon/color aceitam null para limpar.",
        "minProperties": 1,
        "properties": {
          "name": { "type": "string", "maxLength": 100 },
          "slug": { "type": "string", "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$", "maxLength": 120 },
          "description": { "type": "string", "nullable": true },
          "icon": { "type": "string", "maxLength": 50, "nullable": true },
          "color": { "type": "string", "pattern": "^#[0-9a-fA-F]{6}$", "nullable": true },
          "is_active": { "type": "boolean" }
        }
      },
      "Subcategory": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "category_id": { "type": "integer" },
          "name": { "type": "string", "maxLength": 100 },
          "slug": { "type": "string", "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$", "maxLength": 120, "description": "Chave textual estável para uso no front (única dentro da categoria)." },
          "description": { "type": "string", "nullable": true },
          "icon": { "type": "string", "maxLength": 50, "nullable": true, "description": "Identificador nominal do ícone (não é um arquivo)." },
          "is_active": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "SubcategoryCreateInput": {
        "type": "object",
        "required": ["category_id", "name"],
        "properties": {
          "category_id": { "type": "integer", "minimum": 1 },
          "name": { "type": "string", "maxLength": 100 },
          "slug": { "type": "string", "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$", "maxLength": 120, "description": "Opcional. Se omitido, é derivado do nome (sem acentos)." },
          "description": { "type": "string" },
          "icon": { "type": "string", "maxLength": 50 },
          "is_active": { "type": "boolean", "default": true }
        }
      },
      "SubcategoryUpdateInput": {
        "type": "object",
        "description": "Atualização parcial: envie apenas os campos a alterar. Trocar category_id/name/slug revalida a consistência (categoria existente; nome/slug únicos na categoria).",
        "minProperties": 1,
        "properties": {
          "category_id": { "type": "integer", "minimum": 1 },
          "name": { "type": "string", "maxLength": 100 },
          "slug": { "type": "string", "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$", "maxLength": 120 },
          "description": { "type": "string", "nullable": true },
          "icon": { "type": "string", "maxLength": 50, "nullable": true },
          "is_active": { "type": "boolean" }
        }
      },
...
    }
  }
}