# src/controllers/subcategoriesController.js
const subcategoriesService = require('../services/subcategoriesService');
const { slugify } = require('../utils/slugify');

const SLUG_REGEX = /^[a-z0-9]+(?:-[a-z0-9]+)*$/;
const SLUG_MAX = 120;

const list = async (req, res, next) => {
  try {
    const subcategories = await subcategoriesService.listSubcategories();
    res.json(subcategories);
  } catch (err) {
    next(err);
  }
};

const getById = async (req, res, next) => {
  try {
    const id = Number(req.params.id);
    if (!Number.isInteger(id) || id <= 0) {
      return res.status(400).json({ error: 'Invalid id' });
    }

    const subcategory = await subcategoriesService.getSubcategoryById(id);
    if (!subcategory) {
      return res.status(404).json({ error: 'Subcategory not found' });
    }

    res.json(subcategory);
  } catch (err) {
    next(err);
  }
};

const create = async (req, res, next) => {
  try {
    const { category_id, name, slug, description, icon, is_active } = req.body || {};

    if (!Number.isInteger(category_id) || category_id <= 0) {
      return res.status(400).json({ error: 'category_id is required and must be a positive integer' });
    }
    if (typeof name !== 'string' || name.trim().length === 0) {
      return res.status(400).json({ error: 'name is required' });
    }
    if (name.length > 100) {
      return res.status(400).json({ error: 'name must be at most 100 characters' });
    }
    if (icon !== undefined && icon !== null && (typeof icon !== 'string' || icon.length > 50)) {
      return res.status(400).json({ error: 'icon must be a string up to 50 characters' });
    }
    if (is_active !== undefined && typeof is_active !== 'boolean') {
      return res.status(400).json({ error: 'is_active must be a boolean' });
    }

    // slug: usa o informado (validado) ou deriva do nome.
    let finalSlug;
    if (slug !== undefined && slug !== null) {
      if (typeof slug !== 'string' || slug.length > SLUG_MAX || !SLUG_REGEX.test(slug)) {
        return res.status(400).json({ error: `slug must be lowercase alphanumeric words separated by hyphens, up to ${SLUG_MAX} characters` });
      }
      finalSlug = slug;
    } else {
      finalSlug = slugify(name);
      if (!finalSlug) {
        return res.status(400).json({ error: 'could not derive slug from name; provide slug explicitly' });
      }
    }

    const subcategory = await subcategoriesService.createSubcategory({
      category_id,
      name: name.trim(),
      slug: finalSlug,
      description,
      icon,
      is_active,
    });

    res.status(201).json(subcategory);
  } catch (err) {
    if (err.code === 'CATEGORY_NOT_FOUND') {
      return res.status(404).json({ error: err.message });
    }
    if (err.code === 'SUBCATEGORY_NAME_CONFLICT') {
      return res.status(409).json({ error: err.message });
    }
    if (err.code === 'SUBCATEGORY_SLUG_CONFLICT') {
      return res.status(409).json({ error: err.message });
    }
    next(err);
  }
};

const update = async (req, res, next) => {
  try {
    const id = Number(req.params.id);
    if (!Number.isInteger(id) || id <= 0) {
      return res.status(400).json({ error: 'Invalid id' });
    }

    const { category_id, name, slug, description, icon, is_active } = req.body || {};
    const patch = {};

    if (category_id !== undefined) {
      if (!Number.isInteger(category_id) || category_id <= 0) {
        return res.status(400).json({ error: 'category_id must be a positive integer' });
      }
      patch.category_id = category_id;
    }

    if (name !== undefined) {
      if (typeof name !== 'string' || name.trim().length === 0) {
        return res.status(400).json({ error: 'name must be a non-empty string' });
      }
      if (name.length > 100) {
        return res.status(400).json({ error: 'name must be at most 100 characters' });
      }
      patch.name = name.trim();
    }

    if (slug !== undefined) {
      if (typeof slug !== 'string' || slug.length > SLUG_MAX || !SLUG_REGEX.test(slug)) {
        return res.status(400).json({ error: `slug must be lowercase alphanumeric words separated by hyphens, up to ${SLUG_MAX} characters` });
      }
      patch.slug = slug;
    }

    if (description !== undefined) {
      if (description !== null && typeof description !== 'string') {
        return res.status(400).json({ error: 'description must be a string or null' });
      }
      patch.description = description;
    }

    if (icon !== undefined) {
      if (icon !== null && (typeof icon !== 'string' || icon.length > 50)) {
        return res.status(400).json({ error: 'icon must be a string up to 50 characters' });
      }
      patch.icon = icon;
    }

    if (is_active !== undefined) {
      if (typeof is_active !== 'boolean') {
        return res.status(400).json({ error: 'is_active must be a boolean' });
      }
      patch.is_active = is_active;
    }

    if (Object.keys(patch).length === 0) {
      return res.status(400).json({ error: 'No valid fields to update' });
    }

    const subcategory = await subcategoriesService.updateSubcategory(id, patch);
    if (!subcategory) {
      return res.status(404).json({ error: 'Subcategory not found' });
    }

    res.json(subcategory);
  } catch (err) {
    if (err.code === 'CATEGORY_NOT_FOUND') {
      return res.status(404).json({ error: err.message });
    }
    if (err.code === 'SUBCATEGORY_NAME_CONFLICT') {
      return res.status(409).json({ error: err.message });
    }
    if (err.code === 'SUBCATEGORY_SLUG_CONFLICT') {
      return res.status(409).json({ error: err.message });
    }
    next(err);
  }
};

const remove = async (req, res, next) => {
  try {
    const id = Number(req.params.id);
    if (!Number.isInteger(id) || id <= 0) {
      return res.status(400).json({ error: 'Invalid id' });
    }

    const deleted = await subcategoriesService.deleteSubcategory(id);
    if (!deleted) {
      return res.status(404).json({ error: 'Subcategory not found' });
    }

    res.status(204).send();
  } catch (err) {
    if (err.code === 'SUBCATEGORY_IN_USE') {
      return res.status(409).json({ error: err.message });
    }
    next(err);
  }
};

module.exports = { list, getById, create, update, remove };